+++ packages/falcon-ecommerce-uikit/src/Filters/RangeFilter.tsx
@@ -1,0 +1,6 @@
+export type RangeFilterProps = {
+  min: number;
+  max: number;
+  selected: { from: number; to: number };
+  setFilter: (from: string, to: string) => void;
+};
