+++ packages/falcon-ecommerce-uikit/src/Filters/SingleFilter.tsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { ThemedComponentProps } from '@deity/falcon-ui';
+import { FilterOption } from '../Search/types';
+import { FilterItemLayout, SelectedFilterItem } from './FilterItem';
+import { FilterItemsList } from './Filter';
+
+export type SingleFilterProps = {
+  options: FilterOption[];
+  selected?: string;
+  onChange: (value?: string) => void;
+};
+export const SingleFilter: React.SFC<SingleFilterProps & ThemedComponentProps> = ({
+  options,
+  selected,
+  onChange,
+  ...rest
+}) => {
+  const selectedOption = selected !== undefined ? options.find(x => x.value === selected) : undefined;
+
+  return (
+    <FilterItemsList {...rest as any}>
+      {selectedOption && <SelectedFilterItem onClick={() => onChange()}>{selectedOption!.title}</SelectedFilterItem>}
+      {!selectedOption &&
+        options.map(x => (
+          <FilterItemLayout key={x.value} onClick={() => onChange(x.value)}>
+            {x.title} ({x.count})
+          </FilterItemLayout>
+        ))}
+    </FilterItemsList>
+  );
+};
